move HUDManager -> QuickLoot::Behaviors::ActivationBlocker
